# openapi/v0.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '0.1.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Because the Twext compiler does not embed manifest metadata into its
    compiled output (unlike the old Warp/Fluorite `Warp.meta` convention),
    every publish call must submit its manifest fields explicitly as JSON
    alongside the compiled code — TwextHub never parses the JS blob to
    recover metadata.

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they call
    `POST /terms/accept`.

    Bearer tokens come in two kinds. **Sessions** are created by
    `/auth/login` or `/auth/signup`, expire after 7 days, and implicitly
    satisfy every scope. **Automation tokens** are created explicitly via
    `POST /tokens`, are scoped to just `publish` and/or `yank` for CI use
    (e.g. a `twext publish` GitHub Action), and never expire unless the
    caller sets one. Only a session may manage sessions or automation
    tokens, act on `/admin/*` endpoints, or change account settings —
    an automation token calling any of those gets a `403`.
servers:
...
    variables:
...
security:
  - BearerAuth: []
paths:
...
  /@{namespace}/{id}/versions:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
      - $ref: '#/components/parameters/ExtensionIdPath'
    post:
      summary: Publish a new version
      operationId: publishVersion
      tags: [Userspace]
      description: >
        `namespace` in the path must match the authenticated caller (admins
        may publish on behalf of any namespace). A caller's first-ever
        publish is held as `pending`; once an admin approves it, all of
        that caller's later publishes go straight to `published`. Only one
        `pending` version per owner may exist at a time. `manifest.id` must
        equal the `id` path segment, and `manifest.version` must be valid
        SemVer strictly greater than the current highest published version
        for this package (yanked versions still count toward that check).
        Requires the `publish` scope (sessions satisfy this implicitly).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest, code]
              properties:
                manifest: { $ref: '#/components/schemas/Manifest' }
                code:
                  type: string
                  description: The compiled Twext output (a single IIFE JavaScript file), as text.
      responses:
        '201':
          description: Version created (status reflects moderation state)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionInfo' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Namespace mismatch, or the owner already has a version awaiting review
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
components:
  parameters:
...
    NamespacePath:
      name: namespace
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Namespace' }
    ExtensionIdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ExtensionId' }
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
    ExtensionId:
      type: string
      pattern: '^[a-z0-9]{1,64}$'
      description: >
        Matches TurboWarp's own documented rule for extension `id`s: only
        `a-z` and `0-9`, no spaces or special characters. The 64-char cap
        is TwextHub's own storage/URL-length limit, not a TurboWarp
        requirement. twext.yml's `extension.id` field has no pattern
        constraint today; the `twext` CLI's own validation should be
        tightened to match this rather than the other way around.
...
    Manifest:
      type: object
      required: [id, version, license, description]
      description: Mirrors the fields in twext.yml that TwextHub needs to store; not the whole manifest.
      properties:
        id: { $ref: '#/components/schemas/ExtensionId' }
        name: { type: string }
        version: { type: string, description: SemVer. }
        license: { type: string, description: SPDX identifier. }
        description: { type: string }
        author: { type: string }
        color1: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        color2: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        color3: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
...
    VersionInfo:
      type: object
      required: [namespace, id, version, status, createdAt]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        id: { $ref: '#/components/schemas/ExtensionId' }
        version: { type: string }
        status: { type: string, enum: [pending, published, rejected, yanked] }
        name: { type: string }
        license: { type: string }
        description: { type: string }
        author: { type: string }
        createdAt: { type: string, format: date-time }
        publishedAt:
          { type: string, format: date-time, description: Absent while pending or rejected. }
        dist:
          type: object
          description: Absent while pending or rejected.
          properties:
            downloadUrl: { type: string, format: uri }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          { type: string, format: uri, example: 'https://hub.twext.dev/errors/validation-error' }
        title: { type: string, example: 'Validation Error' }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    Conflict:
      description: Resource already exists
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    UnprocessableEntity:
      description: Semantically invalid input (e.g. non-SemVer version, id mismatch)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer